+++ r8ge-core/platform/CMakeLists.txt
@@ -1,0 +1,16 @@
+set(include-headers
+        "include/Time.h"
+)
+
+if(LINUX)
+    set(platform-code
+        "linux-impl/Time.cpp"
+    )
+endif()
+
+add_library(platform
+        ${include-headers}
+        ${platform-code}
+)
+
+target_include_directories(platform PRIVATE "include")
