+++ r8ge-core/platform/include/Time.h
@@ -1,0 +1,14 @@
+#ifndef R8GE_TIME_H
+#define R8GE_TIME_H
+
+namespace r8ge {
+
+    class Time {
+    public:
+        Time();
+    private:
+    };
+
+}
+
+#endif//!R8GE_TIME_H
