+++ r8ge-core/platform/linux-impl/Time.cpp
@@ -1,0 +1,8 @@
+#include "Time.h"
+
+
+namespace r8ge {
+    Time::Time() {
+        
+    }
+}
