--- r8ge-core/platform/linux/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_library(r8ge-core-linux
-)
