+++ r8ge-core/utility/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(utility
+        "StringFormat.h"
+        "StringFormat.cpp"
+)
